+++ src/main/java/com/codegram/cabapi/service/TripBookingService.java
@@ -11,12 +11,12 @@ import com.codegram.cabapi.domain.TripBooking;
  *
  */
 public interface TripBookingService {
-	public TripBooking insertTripBooking(TripBooking tripBooking, Long customerId);
+	public TripBooking insertTripBooking(TripBooking tripBooking, int customerId);
 	public TripBooking updateTripBooking(TripBooking tripBooking);
 	public Iterable<TripBooking> deleteTripBooking(int tripBookingId);
 	public Iterable<TripBooking> findAll();
-	public List<TripBooking> viewAllTripsCustomer(Long customerId);
-	public float calculateBill(Long customerId);
+	public List<TripBooking> viewAllTripsCustomer(int customerId);
+	public float calculateBill(int customerId);
 	public TripBooking viewTripById(int tripBookingId);
 
 }
